Use local node jenkins home instead of remote_fs
Using new_resource.remote_fs prevent the resource to be used from
the master if it happens that the remote_fs does not exist (say if the
master is on Linux and the slave on Windows). Since this is a locally
generated and runned script, using the jenkins home make more sense, and
we are sure it does exist.